fld {pbdML}	R Documentation

Fisher's Linear Discriminant

Description

Compute the 2-class Fisher's linear discriminant either in serial or parallel.

Usage

fld(x, g, checkg = TRUE)

Arguments

x
The data in the form of a matrix or ddmatrix.

g
The group variable in the form of a matrix/vector or a ddmatrix. The values should be 0 and 1 exclusively.

checkg
Logical; should the values of g be confirmed to exclusively be 0's and 1's.

Details

Note that a ddmatrix x and ordinary matrix/vector pairing is possible, but the reverse (matrix/vector x and ddmatrix g) is not.

Value

A list of class 'fld' containing the prior probabilities, group means, w vector, and c scalar. In the distributed case, the priors and c scalar are both global, while the other values are distributed.

Author(s)

Drew Schmidt

References

Duda, R. O., Hart, P. E., & Stork, D. G. (2012). Pattern classification, chapter 5. John Wiley & Sons.

Examples

## Not run: 
x <- matrix(rnorm(30), 10)
g <- sample(0:1, size=10, replace=TRUE)

fld(x, g)

## End(Not run)

[Package pbdML version 0.1-0 Index]